started data extraction to JSON format
added a simple python program which analyses the results from binwalk on a binary and converts it to a JSON formatted file. This needs work; all binwalk outputs seem to be different, so there needs to be a standardised way to convert these into a JSON format. might need to use another tool instead of binwalk, however, binwalk was the best one i found so far. Potentially, could use multiple tools as well